Move the keymask classes to brave.core.util.eve.
